Feature/goals/show goals list (#18)
* Added Progress bar which is showing current goal progress

* Added Goal Progress

+++ features/goals/src/main/java/app/chamich/feature/goals/model/api/IGoal.kt
@@ -8,8 +8,8 @@ interface IGoal {
     val id: Long
     val title: String
     val measuredIn: Int
-    val totalEffort: Long
-    val progress: Long
+    val totalEffort: Int
+    val progress: Int
     val completeData: Long
     val category: Int
     val color: Int
